requirements.txt
dj-database-url
whitenoise
Django
django-pipeline
django-reversion
gunicorn
Markdown
psycopg2
static3
pytz
PyYAML
python-dateutil
django-postgrespool
bleach
django-redis-cache
